+++ go/lib/hpkt/read.go
@@ -12,10 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package hpkt (Host Packet) contains low level primitives for parsing and
-// creating end-host SCION messages.
-//
-// Currently supports SCION/UDP and SCION/SCMP packets.
 package hpkt
 
 import (
@@ -27,13 +23,19 @@ import (
 	"github.com/netsec-ethz/scion/go/lib/util"
 )
 
-// Processing/parsing callback type
-type PktParser func() error
+// ParseScnPkt populates the SCION fields in s with information from b
+func ParseScnPkt(s *spkt.ScnPkt, b common.RawBytes) error {
+	pCtx := newParseCtx(s, b)
+	return pCtx.parse()
+}
 
-// Offsets holds start and end offsets for packet sections
-type Offsets struct {
+// offsets holds start and end offsets for packet sections
+type offsets struct {
 	start, end int
 }
+
+// Processing/parsing callback type
+type pktParser func() error
 
 // parseCtx holds the state for the packet parser
 type parseCtx struct {
@@ -46,48 +48,45 @@ type parseCtx struct {
 	// Helper container for common header fields; also tracks the next
 	// protocol we need to parse
 	cmnHdr *spkt.CmnHdr
+	// Set to true after first E2E extension to guard against misplaced HBH extensions
+	parsedFirstE2E bool
+	// Number of found HBH extensions
+	hbhCounter int
+	// Maximum number of allowed HBH extensions
+	hbhLimit int
+	// Protocol type of next header (L4, HBH class, E2E class)
+	nextHdr common.L4ProtocolType
 
 	// Memorize section start and end offsets for when we need to jump
-	cmnHdrOffsets  Offsets
-	extHdrOffsets  Offsets
-	addrHdrOffsets Offsets
-	fwdPathOffsets Offsets
-	l4HdrOffsets   Offsets
-	pldOffsets     Offsets
+	cmnHdrOffsets  offsets
+	extHdrOffsets  offsets
+	addrHdrOffsets offsets
+	fwdPathOffsets offsets
+	l4HdrOffsets   offsets
+	pldOffsets     offsets
 
 	// Methods for parsing various packet elements; can be overwritten by extensions
 	// FIXME(scrye): when the need arises, these should probably be changed to queues
 	// (e.g., when multiple handlers need to be chained)
-	CmnHdrParser  PktParser
-	HBHExtParser  PktParser
-	E2EExtParser  PktParser
-	AddrHdrParser PktParser
-	FwdPathParser PktParser
-	L4HdrParser   PktParser
-	PldParser     PktParser
-	ChecksumF     PktParser
+	HBHExtParser  pktParser
+	E2EExtParser  pktParser
+	AddrHdrParser pktParser
+	FwdPathParser pktParser
+	L4Parser      pktParser
 }
 
 func newParseCtx(s *spkt.ScnPkt, b common.RawBytes) *parseCtx {
 	pCtx := &parseCtx{
-		s:      s,
-		b:      b,
-		cmnHdr: &spkt.CmnHdr{}}
-	pCtx.CmnHdrParser = pCtx.DefaultCmnHdrParser
+		s:        s,
+		b:        b,
+		cmnHdr:   &spkt.CmnHdr{},
+		hbhLimit: common.ExtnMaxHBH}
 	pCtx.E2EExtParser = pCtx.DefaultE2EExtParser
 	pCtx.HBHExtParser = pCtx.DefaultHBHExtParser
 	pCtx.AddrHdrParser = pCtx.DefaultAddrHdrParser
 	pCtx.FwdPathParser = pCtx.DefaultFwdPathParser
-	pCtx.L4HdrParser = pCtx.DefaultL4HdrParser
-	pCtx.PldParser = pCtx.DefaultPldParser
-	pCtx.ChecksumF = pCtx.DefaultChecksumF
+	pCtx.L4Parser = pCtx.DefaultL4Parser
 	return pCtx
-}
-
-// ParseScnPkt populates the SCION fields in s with information from b
-func ParseScnPkt(s *spkt.ScnPkt, b common.RawBytes) error {
-	pCtx := newParseCtx(s, b)
-	return pCtx.parse()
 }
 
 // parse contains the processing flow
@@ -95,30 +94,48 @@ func (p *parseCtx) parse() error {
 	// A SCION header is parsed in the following order:
 	//  1. Common header
 	//  2. Extension headers, in the order they are placed in the packet.
-	//  Note that extension headers can also overwrite default behavior for
-	//  steps 2-6
 	//  3. Address headers
 	//  4. Forwarding path
 	//  5. L4 header
 	//  6. Payload
+	//
+	// Hop By Hop (HBH) extensions can override 2-6, while End to end (E2E)
+	// extensions can override 5-6.
 
 	if err := p.CmnHdrParser(); err != nil {
 		return common.NewError("Unable to parse common header", "err", err)
 	}
+	p.nextHdr = p.cmnHdr.NextHdr
 
 	// We'll advance the end offset for extensions as we parse them
 	p.extHdrOffsets.start = int(p.cmnHdr.HdrLen * common.LineLen)
 	p.extHdrOffsets.end = p.extHdrOffsets.start
 	// Skip after SCION header
 	p.offset = p.extHdrOffsets.start
+
+	// SCION packets can contain at most 3 HBH extensions, which must appear
+	// immediately after the path header. If an SCMP HBH extension is present,
+	// it must be the first extension and raises the allowed HBH limit to 4.
+	// E2E extensions appear after HBH extensions (if any), or after the path
+	// header.
 ProtoLoop:
 	for {
-		switch p.cmnHdr.NextHdr {
+		switch p.nextHdr {
 		case common.HopByHopClass:
+			p.hbhCounter += 1
+			if p.hbhCounter > p.hbhLimit {
+				return common.NewError("HBH extension limit exceeded", "limit", p.hbhLimit)
+			}
+			if p.parsedFirstE2E {
+				return common.NewError("Invalid placement of HBH extension, must appear before " +
+					"E2E extensions but found after")
+			}
 			if err := p.HBHExtParser(); err != nil {
-				return common.NewError("Unable to parse extension", "err", err)
+				return common.NewError("Unable to parse HBH extension", "err", err)
 			}
 		case common.End2EndClass:
+			// Once we parsed one E2E, we can no longer parse HBHs
+			p.parsedFirstE2E = true
 			if err := p.E2EExtParser(); err != nil {
 				return common.NewError("Unable to parse E2E extension", "err", err)
 			}
@@ -127,7 +144,7 @@ ProtoLoop:
 		case common.L4UDP:
 			break ProtoLoop
 		default:
-			return common.NewError("Unsupported protocol", "proto", p.cmnHdr.NextHdr)
+			return common.NewError("Unsupported protocol", "proto", p.nextHdr)
 		}
 	}
 	// Return to the start of the address header
@@ -141,25 +158,24 @@ ProtoLoop:
 
 	// Jump after extensions
 	p.offset = p.extHdrOffsets.end
-	if err := p.L4HdrParser(); err != nil {
-		return common.NewError("Unable to parse L4 header", "err", err)
-	}
-	if err := p.PldParser(); err != nil {
-		return common.NewError("Unable to parse payload header", "err", err)
-	}
-	if err := p.ChecksumF(); err != nil {
-		return common.NewError("Checksum error", "err", err)
+	if err := p.L4Parser(); err != nil {
+		return common.NewError("Unable to parse L4 content", "err", err)
 	}
 	return nil
 }
 
-func (p *parseCtx) DefaultCmnHdrParser() error {
+func (p *parseCtx) CmnHdrParser() error {
 	p.cmnHdrOffsets.start = p.offset
 	if cerr := p.cmnHdr.Parse(p.b[:spkt.CmnHdrLen]); cerr != nil {
 		return cerr
 	}
 	p.offset += spkt.CmnHdrLen
 	p.cmnHdrOffsets.end = p.offset
+
+	if int(p.cmnHdr.TotalLen) != len(p.b) {
+		return common.NewError("Malformed total packet length", "expected", p.cmnHdr.TotalLen,
+			"actual", len(p.b))
+	}
 	return nil
 }
 
@@ -170,14 +186,24 @@ func (p *parseCtx) DefaultHBHExtParser() error {
 
 	// Parse 3-byte extension header first
 	// We know the type of the next header, so we save it for the protocol loop
-	p.cmnHdr.NextHdr = common.L4ProtocolType(p.b[p.offset])
+	p.nextHdr = common.L4ProtocolType(p.b[p.offset])
 	hdrLen := p.b[p.offset+1]
 	extnType := p.b[p.offset+2]
+	// Advance end of extensions headers offset
+	p.extHdrOffsets.end += int(hdrLen * common.LineLen)
 
 	// Parse the rest of the extension header, depending on extension type
 	switch extnType {
 	case common.ExtnSCMPType.Type:
-		extn, cerr := scmp.ExtnFromRaw(p.b[p.offset+3:])
+		if p.hbhCounter != 1 {
+			// SCMP HBH extensions must come immediately after the path header
+			return common.NewError("Invalid placement of HBH SCMP extension (must be first)",
+				"position", p.hbhCounter, "offset", p.offset)
+		}
+		// SCMP HBH extensions increase the limit of HBH extensions by 1
+		p.hbhLimit += 1
+
+		extn, cerr := scmp.ExtnFromRaw(p.b[p.offset+common.ExtnSubHdrLen : p.extHdrOffsets.end])
 		if cerr != nil {
 			return common.NewError("Unable to parse extension header", "err", cerr)
 		}
@@ -186,9 +212,7 @@ func (p *parseCtx) DefaultHBHExtParser() error {
 		return common.NewError("Unsupported HBH extension type", "type", extnType)
 	}
 
-	// Finished parsing another extension, advance the end offset
-	p.extHdrOffsets.end += int(hdrLen * common.LineLen)
-	p.offset += int(hdrLen * common.LineLen)
+	p.offset = p.extHdrOffsets.end
 	return nil
 }
 
@@ -235,11 +259,11 @@ func (p *parseCtx) DefaultFwdPathParser() error {
 	return nil
 }
 
-func (p *parseCtx) DefaultL4HdrParser() error {
+func (p *parseCtx) DefaultL4Parser() error {
 	var cerr *common.Error
 	p.l4HdrOffsets.start = p.offset
 
-	switch p.cmnHdr.NextHdr {
+	switch p.nextHdr {
 	case common.L4UDP:
 		if p.s.L4, cerr = l4.UDPFromRaw(p.b[p.offset : p.offset+l4.UDPLen]); cerr != nil {
 			return common.NewError("Unable to parse UDP header", "err", cerr)
@@ -250,137 +274,27 @@ func (p *parseCtx) DefaultL4HdrParser() error {
 		}
 	default:
 		return common.NewError("Unsupported NextHdr value", "expected",
-			common.L4UDP, "actual", p.cmnHdr.NextHdr)
+			common.L4UDP, "actual", p.nextHdr)
 	}
-
-	// Enable checksum function
-	p.ChecksumF = func() error {
-		cerr := l4.CheckCSum(p.s.L4, p.b[p.addrHdrOffsets.start:p.addrHdrOffsets.end],
-			p.b[p.pldOffsets.start:p.pldOffsets.end])
-		if cerr != nil {
-			return cerr
-		}
-		return nil
-	}
-
 	p.offset += p.s.L4.L4Len()
 	p.l4HdrOffsets.end = p.offset
-	return nil
-}
 
-func (p *parseCtx) DefaultPldParser() error {
+	// Parse L4 payload
 	p.pldOffsets.start = p.offset
 	pldLen := int(p.cmnHdr.TotalLen) - p.cmnHdr.HdrLenBytes() - p.s.L4.L4Len() -
 		(p.extHdrOffsets.end - p.extHdrOffsets.start)
-	if p.offset+pldLen < len(p.b) {
-		return common.NewError("Incomplete packet, bad payload length",
-			"expected", pldLen, "actual", len(p.b)-p.offset)
+	if cerr = p.s.L4.Validate(pldLen); cerr != nil {
+		return common.NewError("L4 validation failed", "err", cerr)
 	}
 	p.s.Pld = common.RawBytes(p.b[p.offset : p.offset+pldLen])
 	p.offset += pldLen
 	p.pldOffsets.end = p.offset
-	return nil
-}
 
-func (p *parseCtx) DefaultChecksumF() error {
-	return nil
-}
-
-func WriteScnPkt(s *spkt.ScnPkt, b common.RawBytes) (int, error) {
-	var cerr *common.Error
-	offset := 0
-
-	if s.L4.L4Type() != common.L4UDP {
-		return 0, common.NewError("Unsupported protocol", "expected",
-			common.L4UDP, "actual", s.L4.L4Type())
-	}
-	if s.E2EExt != nil {
-		return 0, common.NewError("E2E extensions not supported", "ext", s.E2EExt)
-	}
-	if s.HBHExt != nil {
-		return 0, common.NewError("HBH extensions not supported", "ext", s.HBHExt)
-	}
-
-	// Compute header lengths
-	addrHdrLen := s.DstHost.Size() + s.SrcHost.Size() + 2*addr.IABytes
-	addrPad := util.CalcPadding(addrHdrLen, common.LineLen)
-	addrHdrLen += addrPad
-	pathHdrLen := 0
-	if s.Path != nil {
-		pathHdrLen = len(s.Path.Raw)
-	}
-	scionHdrLen := spkt.CmnHdrLen + addrHdrLen + pathHdrLen
-	pktLen := scionHdrLen + s.L4.L4Len() + s.Pld.Len()
-	if len(b) < pktLen {
-		return 0, common.NewError("Buffer too small", "expected", pktLen,
-			"actual", len(b))
-	}
-
-	// Compute preliminary common header, but do not write it to the packet yet
-	cmnHdr := spkt.CmnHdr{}
-	cmnHdr.Ver = spkt.SCIONVersion
-	cmnHdr.DstType = s.DstHost.Type()
-	cmnHdr.SrcType = s.SrcHost.Type()
-	cmnHdr.TotalLen = uint16(pktLen)
-	cmnHdr.HdrLen = uint8(scionHdrLen / common.LineLen)
-	cmnHdr.CurrInfoF = 0 // Updated later if necessary
-	cmnHdr.CurrHopF = 0  // Updated later if necessary
-	cmnHdr.NextHdr = s.L4.L4Type()
-	offset += spkt.CmnHdrLen
-
-	// Address header
-	addrSlice := b[offset : offset+addrHdrLen]
-	s.DstIA.Write(b[offset:])
-	offset += addr.IABytes
-	s.SrcIA.Write(b[offset:])
-	offset += addr.IABytes
-	// addr.HostAddr.Pack() is zero-copy, use it directly
-	offset += copy(b[offset:], s.DstHost.Pack())
-	offset += copy(b[offset:], s.SrcHost.Pack())
-	// Zero memory padding
-	zeroMemory(b[offset : offset+addrPad])
-	offset += addrPad
-
-	// Forwarding Path
-	if s.Path != nil {
-		cmnHdr.CurrInfoF = uint8((offset + s.Path.InfOff) / common.LineLen)
-		cmnHdr.CurrHopF = uint8((offset + s.Path.HopOff) / common.LineLen)
-		offset += copy(b[offset:], s.Path.Raw)
-	}
-
-	// Write the common header at the start of the buffer
-	cmnHdr.Write(b)
-
-	// Don't write L4 yet
-	l4Slice := b[offset : offset+s.L4.L4Len()]
-	offset += s.L4.L4Len()
-
-	// Payload
-	pldSlice := b[offset : offset+s.Pld.Len()]
-	s.Pld.WritePld(b[offset:])
-	offset += s.Pld.Len()
-
-	// SCION/UDP Header
-	cerr = l4.SetCSum(s.L4, addrSlice, pldSlice)
+	// Run checksum function
+	cerr = l4.CheckCSum(p.s.L4, p.b[p.addrHdrOffsets.start:p.addrHdrOffsets.end],
+		p.b[p.pldOffsets.start:p.pldOffsets.end])
 	if cerr != nil {
-		return 0, common.NewError("Unable to compute checksum", "err", cerr)
+		return common.NewError("Checksum failed", "err", cerr)
 	}
-	s.L4.Write(l4Slice)
-
-	return offset, nil
-}
-
-func isZeroMemory(b common.RawBytes) (int, bool) {
-	for i := range b {
-		if b[i] != 0 {
-			return i, false
-		}
-	}
-	return 0, true
-}
-
-func zeroMemory(b common.RawBytes) {
-	for i := range b {
-		b[i] = 0
-	}
+	return nil
 }
