+++ go/lib/hpkt/write.go
@@ -1,0 +1,126 @@
+// Copyright 2017 ETH Zurich
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package hpkt (Host Packet) contains low level primitives for parsing and
+// creating end-host SCION messages.
+//
+// Currently supports SCION/UDP and SCION/SCMP packets.
+package hpkt
+
+import (
+	"github.com/netsec-ethz/scion/go/lib/addr"
+	"github.com/netsec-ethz/scion/go/lib/common"
+	"github.com/netsec-ethz/scion/go/lib/l4"
+	"github.com/netsec-ethz/scion/go/lib/spkt"
+	"github.com/netsec-ethz/scion/go/lib/util"
+)
+
+func WriteScnPkt(s *spkt.ScnPkt, b common.RawBytes) (int, error) {
+	var cerr *common.Error
+	offset := 0
+
+	if s.L4.L4Type() != common.L4UDP {
+		return 0, common.NewError("Unsupported protocol", "expected",
+			common.L4UDP, "actual", s.L4.L4Type())
+	}
+	if s.E2EExt != nil {
+		return 0, common.NewError("E2E extensions not supported", "ext", s.E2EExt)
+	}
+	if s.HBHExt != nil {
+		return 0, common.NewError("HBH extensions not supported", "ext", s.HBHExt)
+	}
+
+	// Compute header lengths
+	addrHdrLen := s.DstHost.Size() + s.SrcHost.Size() + 2*addr.IABytes
+	addrPad := util.CalcPadding(addrHdrLen, common.LineLen)
+	addrHdrLen += addrPad
+	pathHdrLen := 0
+	if s.Path != nil {
+		pathHdrLen = len(s.Path.Raw)
+	}
+	scionHdrLen := spkt.CmnHdrLen + addrHdrLen + pathHdrLen
+	pktLen := scionHdrLen + s.L4.L4Len() + s.Pld.Len()
+	if len(b) < pktLen {
+		return 0, common.NewError("Buffer too small", "expected", pktLen,
+			"actual", len(b))
+	}
+
+	// Compute preliminary common header, but do not write it to the packet yet
+	cmnHdr := spkt.CmnHdr{}
+	cmnHdr.Ver = spkt.SCIONVersion
+	cmnHdr.DstType = s.DstHost.Type()
+	cmnHdr.SrcType = s.SrcHost.Type()
+	cmnHdr.TotalLen = uint16(pktLen)
+	cmnHdr.HdrLen = uint8(scionHdrLen / common.LineLen)
+	cmnHdr.CurrInfoF = 0 // Updated later if necessary
+	cmnHdr.CurrHopF = 0  // Updated later if necessary
+	cmnHdr.NextHdr = s.L4.L4Type()
+	offset += spkt.CmnHdrLen
+
+	// Address header
+	addrSlice := b[offset : offset+addrHdrLen]
+	s.DstIA.Write(b[offset:])
+	offset += addr.IABytes
+	s.SrcIA.Write(b[offset:])
+	offset += addr.IABytes
+	// addr.HostAddr.Pack() is zero-copy, use it directly
+	offset += copy(b[offset:], s.DstHost.Pack())
+	offset += copy(b[offset:], s.SrcHost.Pack())
+	// Zero memory padding
+	zeroMemory(b[offset : offset+addrPad])
+	offset += addrPad
+
+	// Forwarding Path
+	if s.Path != nil {
+		cmnHdr.CurrInfoF = uint8((offset + s.Path.InfOff) / common.LineLen)
+		cmnHdr.CurrHopF = uint8((offset + s.Path.HopOff) / common.LineLen)
+		offset += copy(b[offset:], s.Path.Raw)
+	}
+
+	// Write the common header at the start of the buffer
+	cmnHdr.Write(b)
+
+	// Don't write L4 yet
+	l4Slice := b[offset : offset+s.L4.L4Len()]
+	offset += s.L4.L4Len()
+
+	// Payload
+	pldSlice := b[offset : offset+s.Pld.Len()]
+	s.Pld.WritePld(b[offset:])
+	offset += s.Pld.Len()
+
+	// SCION/UDP Header
+	cerr = l4.SetCSum(s.L4, addrSlice, pldSlice)
+	if cerr != nil {
+		return 0, common.NewError("Unable to compute checksum", "err", cerr)
+	}
+	s.L4.Write(l4Slice)
+
+	return offset, nil
+}
+
+func isZeroMemory(b common.RawBytes) (int, bool) {
+	for i := range b {
+		if b[i] != 0 {
+			return i, false
+		}
+	}
+	return 0, true
+}
+
+func zeroMemory(b common.RawBytes) {
+	for i := range b {
+		b[i] = 0
+	}
+}
